refactor(api): extract menus + redirects + configs domains (#35)
